Reorganize classes into modules.
