+++ app/src/main/java/com/amiel/moviecenter/DB/LocalDB/Converters.java
@@ -1,4 +1,4 @@
-package com.amiel.moviecenter.DB;
+package com.amiel.moviecenter.DB.LocalDB;
 
 import androidx.room.TypeConverter;
